requirements.txt
Flask~=3.1.0
requests~=2.32.3
firebase-admin==5.3.0
python-dotenv==0.19.2
gunicorn==20.1.0